Added EmDocument
EmDocument is a new embedded document system that stores basically a
fancy DictProperty with type checking and conversion.

Forgot to mention for the previous commit, there's a change where it's
backward incompatible due to the fact that the unique bucket name has
changed from _<bucketname>_ul_<propertyname> to
_<classname>_ul_<propertyname>

A script will be made to shuffle the data in the comming commits under
the upgrade_scripts folder.