Fixed another couple bugs relating to referencing
One bug is the reference_class of collection_name properties are set to
DocumentMetaclass, which is fixed now.

Another bug is the fact that cleanupDataFromDatabase don't look at
reference properties, only regular properties.

The last bug is that loading will go into infinite recursion in some
cases of collection_name