Changed default value behaviour
Now, any type where the defaultValue() returns not None, the
initialization of Document or EmDocument will first write the value
returned by it to the associated attribute. Since __getattr__ returns
None if a property is not present but in the attributes defined, it's as
if all defaultValue is set to the data first.