Fixed a bug in deleted references
Before, if we were to delete a reference in a reference property or a
linked document and there's a collection_name, it would not delete the
object from the collection_name. This is now fixed with a relatively
dirty hack.

I'm hoping that eventually I'll be able to track changes instead of just
going through everything.. But riak doesn't even support that atm.