Bug fixes and behaviour changes
There was a default handling bug in EmDocument that prevented default
functions from regenerating a new default value on every initialization.

Another bug was preventing the attribute of "key" from being used in
EmDocuments.

Major Behaviour Changes:

Before, we could initialize an infinite amount of objects with the same
key. This is now restricted. There will only be 1 object per key, so any
changes to 1 object WILL affect the other one. Unless needed, there will
be no way to have 2 objects with the same key in memory.

The reason that this is done, is so that in a 3 way reference, it will
not cause synchronization issue when saving one first and the other one
second, which I have ran into.