Rewrote a good portion of Riakkit.
Got rid of LinkedDocument, got rid of _meta["_references"] and all the
unnecessary code along with it. Also attempting to fix an unique bug
which wouldn't remove the unique entry when del an attribute. However
currently the test is spitting out a 405 (from riak). Not sure why
that's happening yet.

No efficiency boost yet. Currently exploring the option of forking
python-riak.. Also.. no time due to the fact that I'm working on a much
larger project with flask and riakkit.