Added load as an inverse for toRiakObj
SimpleDocument.load is in many ways very similar to Document.load,
however it's much less flexible.

The implementation of SimpleDocument is largely complete. As long as
there's nothing else special, no other feature is likely to be added to
SimpleDocument to avoid bloat for the low level libraries.

There is the possibilities of deriving RiakObject. That depends on what
Basho wants to do with riak-python-client