Dynamic load of references
Before we had the problem where if you have a network of references, it
will load an extremely high amount of documents from the database,
taking very long.

However, this issue has been patched with dynamic loading of references.
This approach will only load an object at __getattr__.

The unittest shows that the speed has been decreased. I cannot confirm
that if the speed actually increased in practise or not at this moment
(working on that, will figure out before merging to master), I suspect
it will as all unittest does .delete at the end. This operation will
load the references if there's collection_name. Therefore it may take
longer because now we load more

However in practise we delete less, so it should be okay.