Rewrite is in progress....
Complete redesign. This time we're deprecating most features that's
meaningless (such as unique, collection_name, and such silliness). A new
design will be in place soon.

In the backend there's a fair bit of change as we are not going to rely
on `RiakObject` to put into the database anymore. Instead, `Document` will
be compatible with `RiakObject` and it can be taken by `RiakClient`.

Still need to investigate if it is possible to make `Document` compatible
with the `RiakClient` get operations. That will be very nice as a lot of
time will be wasted if we go through `RiakObject` as we already store
the data locally. This does mean that there will be some
attributes/methods exposed for riak-python-client that shouldn't be
called by developers, which is not a big deal as warning messages will
be written in the documentation.

Also adopted PEP8, with a 2 space indentation modification.

Let the fun begin.